Initial commit of assg04 template
Initial commit of empty assignment 04 template.
This assignment is on doing streaming file I/O
using the C++ streaming library, including
outputting to a file stream and using I/O
manipulators to format the output.

This template should be complete, barring any
bugs or typos that need to be corrected.